moved to vite 4.4.9 started typescript conversion on controllers, models
