+++ index.html
@@ -13,6 +13,7 @@
 <body>
 
   <div id="root"></div>
+  <script type="module" src="/src/index.tsx"></script>
 
 </body>
