+++ server/models/sessionModel.ts
@@ -1,4 +1,4 @@
-const mongoose = require('mongoose');
+import mongoose from 'mongoose';
 const Schema = mongoose.Schema;
 
 const sessionSchema = new Schema({
@@ -8,4 +8,4 @@ const sessionSchema = new Schema({
 
 });
 
-module.exports = mongoose.model('Session', sessionSchema);
+export default mongoose.model('Session', sessionSchema);
