+++ server/models/userModel.ts
@@ -1,4 +1,5 @@
-const mongoose = require('mongoose');
+import mongoose from 'mongoose';
+
 const Schema = mongoose.Schema;
 
 const userSchema = new Schema({
@@ -21,4 +22,5 @@ const userSchema = new Schema({
     keepMeLoggedIn:{type: String, unique:true}
 });
 
-module.exports = mongoose.model('User', userSchema);
+export default mongoose.model('User', userSchema);
+
