+++ src/App.tsx
@@ -9,7 +9,7 @@ import SignIn from './components/Login';
 import SignUp from './components/Signup';
 import Splash from './components/Splash';
 import SongPage from './components/SongPage';
-import TopTracks from './components/TopTracks';
+// import TopTracks from './components/TopTracks';
 import AccountSettings from './components/AccountSettings';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
 import backgroundImg from './assets/Musicplugbg.jpg';
