+++ src/components/DbScale.tsx
