+++ src/components/Splash.tsx
