+++ tsconfig.json
@@ -1,0 +1,34 @@
+{
+    "compilerOptions": {
+      "noEmitOnError": false,
+        "module": "commonjs",
+        "esModuleInterop": true,
+        "target": "es6",
+        "moduleResolution": "node",
+        "lib": [
+            "dom",
+            "dom.iterable",
+            "esnext"
+        ],
+        "allowJs": false,
+        "skipLibCheck": true,
+        "allowSyntheticDefaultImports": true,
+        "strict": false,
+        "noImplicitAny": false,
+        "forceConsistentCasingInFileNames": true,
+        "resolveJsonModule": true,
+        "isolatedModules": true,
+        "noEmit": true,
+        "noFallthroughCasesInSwitch": true,
+        "jsx": "react-jsx",
+        "types": [
+            "vite/client",
+            "vite-plugin-svgr/client",
+            "node"
+        ]
+    },
+    "include": [
+        "src",
+        "server/**/*"
+    ]
+}
