+++ vite.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react-swc'
+
+export default defineConfig({
+  base: '/',
+  plugins: [react()],
+  server: {
+    proxy: {
+      '/api': 'http://localhost:4000' // adjust this according to your API routes
+    }
+  }
+})
